refactor: 암호화가 변경되면 주입하는 암호화에 따라 유연하게 수정할 수 있도록 interface 를 만들어 구현하는 형태로 변경
