+++ src/main/java/com/flab/demo/hash/Hash.java
@@ -1,0 +1,5 @@
+package com.flab.demo.hash;
+
+public interface Hash {
+    String digest(String value, String algorithm);
+}
